Move TimeType to own file
